+++ utbot-framework/src/main/kotlin/org/utbot/tests/infrastructure/UtValueTestCaseChecker.kt
@@ -2769,21 +2769,3 @@ inline fun <reified T> withSettingsFromTestFrameworkConfiguration(
         TestCodeGeneratorPipeline.currentTestFrameworkConfiguration = previousConfig
     }
 }
-
-
-
-
-
-
-/**
- * Run [block] with disabled sandbox in the concrete executor
- */
-inline fun <reified T> withoutSandbox(block: () -> T): T {
-    val prev = UtSettings.disableSandbox
-    UtSettings.disableSandbox = true
-    try {
-        return block()
-    } finally {
-        UtSettings.disableSandbox = prev
-    }
-}
